Support ajax for CartController

+++ app/Http/Controllers/CartController.php
@@ -10,11 +10,15 @@ use Illuminate\Support\Facades\Auth;
 
 class CartController extends Controller
 {
-    public function index()
+    public function index(Request $request)
     {
         /** @var User $user */
         $user = Auth::user();
         $cart = $user->cart;
+
+        if($request->ajax()) {
+            return response()->json(["cart" => $cart]);
+        }
 
         return view(
             'user.cart',
@@ -33,6 +37,9 @@ class CartController extends Controller
         $product = Product::find($request->get('product_id'));
         $quantity = $request->get('quantity');
         if (!$product) {
+            if($request->ajax()) {
+                return response()->json(["message" => "Product not found."], 400);
+            }
             return redirect()->back()->with('error', 'Product not found.');
         }
 
@@ -41,10 +48,17 @@ class CartController extends Controller
 
         if($quantity <= 0) {
             $user->cart()->detach($product);
+            if($request->ajax()) {
+                return response()->json(["message" => "Product removed from cart."]);
+            }
             return redirect()->back()->with('message', 'Product removed from cart.');
         }
 
         $user->cart()->attach($product, ['quantity' => $request->get('quantity')]);
+
+        if($request->ajax()) {
+            return response()->json(["message" => "Product added to cart."]);
+        }
 
         return redirect()->back()->with('message', 'Product added to cart.');
     }
@@ -57,10 +71,20 @@ class CartController extends Controller
 
         // check product_id valid
         $product = Product::find($request->get('product_id'));
+        if(!$product) {
+            if($request->ajax()) {
+                return response()->json(["message" => "Product not found."], 400);
+            }
+            return redirect()->back()->with('error', 'Product not found.');
+        }
 
         /** @var User $user */
         $user = Auth::user();
         $user->cart()->detach($product);
+
+        if($request->ajax()) {
+            return response()->json(["message" => "Product removed from cart."]);
+        }
         return redirect()->back()->with('message', 'Product removed from cart.');
     }
 }
